Added department subdocument

+++ src/app/read/read.component.ts
@@ -1,10 +1,9 @@
-import { ChangeDetectionStrategy, Component, OnInit, Output, Input } from '@angular/core';
-import Employee from '../Employee';
+import { ChangeDetectionStrategy, Component, OnInit, ViewContainerRef } from '@angular/core';
 import { PersonService } from '../person.service';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { UpdateComponent} from '../update/update.component';
+// import {ToastrManager} from 'ng6-toastr-notifications';
 
 @Component({
   selector: 'app-read',
@@ -12,72 +11,87 @@ import { UpdateComponent} from '../update/update.component';
   styleUrls: ['./read.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
+
 export class ReadComponent implements OnInit {
 
-  // employeeObjects: any;
+  employeeObjects: any;
   // public filter: string = '';
-  employeeObjects: Observable<string[]>;
+  // employeeObjects: Observable<string[]>;
   page: number = 1;
   count: number = 0;
   loading: boolean;
-  departmnet: Observable<string[]>;
-  sort: string='';
-  order: string='';
+  drpdnDepartment: string;
+  sort: string = '';
+  order: string = '';
+  departments: Observable<string[]>;
 
   filterForm: FormGroup;
 
   ngOnInit() {
     this.getPage(1);
-    // this.getDepartment();
+    this.getDepartment();
   }
 
-  sortData(sort?: string, fname?: string, fgender?: string){
-    // console.log("ok SORT:",, '\n Filter', fname, "Gender", fgender);
-    if(this.order === sort) sort = '-'+sort;
+  sortData(sort?: string, fname?: string, drpdnDepartment?:string, fgender?: string) {
+    console.log("param",drpdnDepartment);
+    if (this.order === sort) sort = '-' + sort;
     this.order = sort;
-    if((fname || fgender) && !sort) var page = 1;
+    if ((fname || fgender || drpdnDepartment) && !sort) var page = 1;
     else page = this.page;
-    this.getPage(page, sort, fname, fgender);
+    this.getPage(page, sort, fname, drpdnDepartment, fgender);
   }
 
-  getPage(page: number=1, sort?: string, fname?: string, fgender?: string) {
+  getPage(page: number = 1, sort?: string, fname?: string,drpdnDepartment?: string, fgender?: string) {
     this.loading = true;
-    console.log("1", sort, fname, fgender);
-    // console.log("PAGE:: ",page);
-    this.ps.getEmployees(page, sort, fname, fgender).then((result) => {
+    page = page || 1;
+    console.log("param",drpdnDepartment);
+    this.ps.getEmployees(page, sort, fname,drpdnDepartment, fgender).then((result) => {
       console.log("RESULT", result['data']);
-      console.log("result", result);
+      // console.log("result", result);
       this.employeeObjects = result['data'];
+      // console.log(this.employeeObjects);
       this.count = result['count'];
       this.page = result['page'];
       this.loading = false;
     });
   }
 
-  getDepartment(){
+  getDepartment() {
+    this.ps.getDepartment().then((result) => {
+      this.departments = result['departments'];
+      console.log('dept',this.departments,'\t', result);
 
+    })
   }
 
-  constructor(private ps: PersonService, private router: Router, private fb: FormBuilder ) {
+  constructor(private ps: PersonService, private router: Router, private fb: FormBuilder) {
     this.createForm();
   }
 
-  createForm(){
+  createForm() {
     this.filterForm = this.fb.group({
       sort: ['', Validators.required],
       fname: ['', Validators.required],
       fgender: ['', Validators.required],
-      fdepartment: ['', Validators.required]
+      fdepartment: ['', Validators.required],
+      drpdnDepartment: ['']
     });
   }
 
-  deleteEmployee(id) {
-    this.ps.deleteEmployee(id).then(res => {
-      console.log('deleted');
-      //location.reload();
-      //this.router.navigate(['employee']);
+  deleteEmployee(id,ind) {
+    var result =confirm("Are you sure you want to delete?");
+    if(result) {
+      this.ps.deleteEmployee(id).then(res => {
+      // this.showSuccess('Deleted Successfully');
+      var index = this.employeeObjects.indexOf(id);
+      console.log(index);
+      this.employeeObjects.splice(ind, 1);
     });
-
   }
+  }
+
+  // showSuccess(msg?: string){
+  //   // this.toastr.successToastr(msg);
+  // }
 
 }
